Enhance estimate creation process with improved validation and logging; update duration type to decimal for consistency across DTOs and frontend components.

+++ Mechanical_workshop/DTOs/EstimateCreateDto.cs
@@ -82,8 +82,8 @@ namespace Mechanical_workshop.Dtos
         [Required]
         public string Description { get; set; } = string.Empty;
 
-        [Range(0, int.MaxValue)]
-        public int Duration { get; set; }
+        [Range(0, double.MaxValue)]
+        public decimal Duration { get; set; }
 
         [Range(0, double.MaxValue)]
         public decimal LaborRate { get; set; }
